refactor(@vtmn/vue): file structure
