+++ packages/showcases/vue/.storybook/main.js
@@ -1,5 +1,9 @@
 module.exports = {
-  stories: ['../stories/**/*.stories.mdx', '../stories/**/*.stories.@(js|jsx|ts|tsx)'],
+  stories: [
+    '../stories/**/*.stories.@(js|jsx|ts|tsx)',
+    '../stories/guidelines/**/*.stories.mdx',
+    '../stories/components/**/*.stories.mdx',
+  ],
   addons: [
     '@storybook/addon-essentials',
     '@storybook/addon-a11y',
