+++ packages/sources/vue/src/components/index.ts
@@ -1,15 +1,15 @@
-import VtmnButton from './VtmnButton';
-import VtmnCheckbox from './VtmnCheckbox';
-import VtmnLink from './VtmnLink';
-import VtmnRadioButton from './VtmnRadioButton';
-import VtmnTextInput from './VtmnTextInput';
-import VtmnToggle from './VtmnToggle';
+import VtmnButton from './actions/VtmnButton';
+import VtmnLink from './actions/VtmnLink';
+import VtmnTextInput from './forms/VtmnTextInput';
+import VtmnCheckbox from './selection-controls/VtmnCheckbox';
+import VtmnRadioButton from './selection-controls/VtmnRadioButton';
+import VtmnToggle from './selection-controls/VtmnToggle';
 
 export {
   VtmnButton,
-  VtmnCheckbox,
   VtmnLink,
-  VtmnRadioButton,
   VtmnTextInput,
+  VtmnCheckbox,
+  VtmnRadioButton,
   VtmnToggle,
 };
