+++ packages/sources/vue/src/guidelines/iconography/VtmnIcon/VtmnIcon.vue
