+++ packages/sources/vue/src/guidelines/iconography/VtmnIcon/index.ts
