+++ packages/sources/vue/src/guidelines/iconography/VtmnIcon/types.ts
