+++ packages/sources/vue/src/guidelines/index.ts
@@ -1,0 +1,3 @@
+import VtmnIcon from './iconography/VtmnIcon';
+
+export { VtmnIcon };
